+++ torch/ao/quantization/quantizer/xnnpack_quantizer.py
@@ -4,7 +4,6 @@ import copy
 import functools
 import itertools
 
-import operator
 from typing import Any, Callable, Dict, List, Optional, Set
 
 import torch
@@ -20,19 +19,7 @@ from torch.ao.quantization.observer import (
     PlaceholderObserver,
 )
 
-from torch.ao.quantization.pt2e.graph_utils import find_sequential_partitions
 from torch.ao.quantization.qconfig import _ObserverOrFakeQuantizeConstructor
-
-from torch.ao.quantization.quantizer.utils import (
-    _annotate_input_qspec_map,
-    _annotate_output_qspec,
-    _is_sym_size_node,
-    _node_only_used_for_sym_size,
-    get_bias_qspec,
-    get_input_act_qspec,
-    get_output_act_qspec,
-    get_weight_qspec,
-)
 
 from torch.fx import Node
 from torch.fx.passes.utils.source_matcher_utils import get_source_partitions
@@ -43,20 +30,18 @@ from .quantizer import (
     QuantizationAnnotation,
     QuantizationConfig,
     QuantizationSpec,
-    QuantizationSpecBase,
     Quantizer,
     SharedQuantizationSpec,
 )
 
-from .xnnpack_quantizer_utils import (
-    _OP_TO_ANNOTATOR,
-)
+from .xnnpack_quantizer_utils import _is_annotated, _OP_TO_ANNOTATOR
 
 
 __all__ = [
     "XNNPACKQuantizer",
     "get_symmetric_quantization_config",
 ]
+
 
 def _get_dynamo_graph(function: Callable, inputs) -> torch.fx.Graph:
     gm, _ = torchdynamo.export(function, aten_graph=True)(*inputs)
@@ -202,21 +187,6 @@ def get_symmetric_quantization_config(
 
 def _get_supported_config_and_operators() -> List[OperatorConfig]:
     return _get_supported_symmetric_config_and_operators()
-
-
-def _is_annotated(nodes: List[Node]):
-    """
-    Given a list of nodes (that represents an operator pattern),
-    check if any of the node is annotated, return True if any of the node
-    is annotated, otherwise return False
-    """
-    annotated = False
-    for node in nodes:
-        annotated = annotated or (
-            "quantization_annotation" in node.meta
-            and node.meta["quantization_annotation"]._annotated
-        )
-    return annotated
 
 
 def _get_module_name_filter(module_name: str):
@@ -369,7 +339,7 @@ class XNNPACKQuantizer(Quantizer):
         self._annotate_hardtanh(model, config, filter_fn)
         self._annotate_mean(model, config, filter_fn)
         self._annotate_adaptive_avg_pool2d(model, config, filter_fn)
-        self._annotate_gru(model, config, filter_fn)
+        self._annotate_gru_io_only(model, config, filter_fn)
         return model
 
     def _annotate_for_static_quantization_config(
@@ -419,53 +389,9 @@ class XNNPACKQuantizer(Quantizer):
         filter_fn: Optional[Callable[[Node], bool]] = None,
     ) -> None:
         """
-        Find Conv2d + batchnorm parititions
         Note: This is only used for QAT. In PTQ, batchnorm should already be fused into the conv.
         """
-        fused_partitions = find_sequential_partitions(
-            gm, [torch.nn.Conv2d, torch.nn.BatchNorm2d], filter_fn
-        )
-        for fused_partition in fused_partitions:
-            conv_partition, bn_partition = fused_partition
-            if len(conv_partition.output_nodes) > 1:
-                raise ValueError("conv partition has more than one output node")
-            conv_node = conv_partition.output_nodes[0]
-            conv_node_users = list(conv_node.users.keys())
-            if len(conv_node_users) > 1:
-                raise ValueError(
-                    "Conv node must be consumed by BN only for it to be fusable."
-                )
-            if len(bn_partition.output_nodes) > 1:
-                raise ValueError("BatchNorm partition has more than one output node")
-            bn_output_node = bn_partition.output_nodes[0]
-
-            if _is_annotated([bn_output_node, conv_node]):
-                continue
-
-            input_qspec_map = {}
-            input_act = conv_node.args[0]
-            assert isinstance(input_act, Node)
-            input_qspec_map[input_act] = get_input_act_qspec(quantization_config)
-
-            weight = conv_node.args[1]
-            assert isinstance(weight, Node)
-            input_qspec_map[weight] = get_weight_qspec(quantization_config)
-
-            bias = conv_node.args[2]
-            if isinstance(bias, Node):
-                input_qspec_map[bias] = get_bias_qspec(quantization_config)
-
-            conv_node.meta["quantization_annotation"] = QuantizationAnnotation(
-                input_qspec_map=input_qspec_map, _annotated=True
-            )
-
-            bn_output_node.meta["quantization_annotation"] = QuantizationAnnotation(
-                output_qspec=get_output_act_qspec(quantization_config),  # type: ignore[arg-type]
-                _annotated=True,
-            )
-            nodes_to_mark_annotated = list(conv_partition.nodes)
-            nodes_to_mark_annotated.extend(list(bn_partition.nodes))
-            _mark_nodes_as_annotated(nodes_to_mark_annotated)
+        return _OP_TO_ANNOTATOR["conv2d_bn"](gm, quantization_config, filter_fn)
 
     def _annotate_conv2d_bn_relu(
         self,
@@ -474,57 +400,9 @@ class XNNPACKQuantizer(Quantizer):
         filter_fn: Optional[Callable[[Node], bool]] = None,
     ) -> None:
         """
-        Find Conv2d + batchnorm + relu parititions
         Note: This is only used for QAT. In PTQ, batchnorm should already be fused into the conv.
         """
-        fused_partitions = find_sequential_partitions(
-            gm, [torch.nn.Conv2d, torch.nn.BatchNorm2d, torch.nn.ReLU], filter_fn
-        )
-        for fused_partition in fused_partitions:
-            conv_partition, bn_partition, relu_partition = fused_partition
-            if len(relu_partition.output_nodes) > 1:
-                raise ValueError("Relu partition has more than one output node")
-            relu_node = relu_partition.output_nodes[0]
-            if len(conv_partition.output_nodes) > 1:
-                raise ValueError("conv partition has more than one output node")
-            conv_node = conv_partition.output_nodes[0]
-            conv_node_users = list(conv_node.users.keys())
-            if len(conv_node_users) > 1:
-                raise ValueError(
-                    "Conv node must be consumed by BN only for it to be fusable."
-                )
-            if len(bn_partition.output_nodes) > 1:
-                raise ValueError("BatchNorm partition has more than one output node")
-            bn_output_node = bn_partition.output_nodes[0]
-
-            if _is_annotated([relu_node, bn_output_node, conv_node]):
-                continue
-
-            input_qspec_map = {}
-            input_act = conv_node.args[0]
-            assert isinstance(input_act, Node)
-            input_qspec_map[input_act] = get_input_act_qspec(quantization_config)
-
-            weight = conv_node.args[1]
-            assert isinstance(weight, Node)
-            input_qspec_map[weight] = get_weight_qspec(quantization_config)
-
-            bias = conv_node.args[2]
-            if isinstance(bias, Node):
-                input_qspec_map[bias] = get_bias_qspec(quantization_config)
-
-            conv_node.meta["quantization_annotation"] = QuantizationAnnotation(
-                input_qspec_map=input_qspec_map, _annotated=True
-            )
-
-            relu_node.meta["quantization_annotation"] = QuantizationAnnotation(
-                output_qspec=get_output_act_qspec(quantization_config),  # type: ignore[arg-type]
-                _annotated=True,
-            )
-            nodes_to_mark_annotated = list(conv_partition.nodes)
-            nodes_to_mark_annotated.extend(list(bn_partition.nodes))
-            nodes_to_mark_annotated.extend(list(relu_partition.nodes))
-            _mark_nodes_as_annotated(nodes_to_mark_annotated)
+        return _OP_TO_ANNOTATOR["conv2d_bn_relu"](gm, quantization_config, filter_fn)
 
     def _annotate_conv2d_relu(
         self,
@@ -532,54 +410,7 @@ class XNNPACKQuantizer(Quantizer):
         quantization_config: QuantizationConfig,
         filter_fn: Optional[Callable[[Node], bool]] = None,
     ) -> None:
-        fused_partitions = find_sequential_partitions(
-            gm, [torch.nn.Conv2d, torch.nn.ReLU], filter_fn
-        )
-        for fused_partition in fused_partitions:
-            conv_partition, relu_partition = fused_partition
-            if len(relu_partition.output_nodes) > 1:
-                raise ValueError("Relu partition has more than one output node")
-            relu_node = relu_partition.output_nodes[0]
-            if len(conv_partition.output_nodes) > 1:
-                raise ValueError("conv partition has more than one output node")
-            conv_node = conv_partition.output_nodes[0]
-
-            if not isinstance(conv_node, Node):
-                raise ValueError(f"{conv_node} is not a Node")
-            if (
-                conv_node.op != "call_function"
-                or conv_node.target != torch.ops.aten.convolution.default
-            ):
-                raise ValueError(f"{conv_node} is not an aten conv2d operator")
-            if relu_node.op != "call_function" or relu_node.target not in [
-                torch.ops.aten.relu.default,
-                torch.ops.aten.relu_.default,
-            ]:
-                raise ValueError(f"{relu_node} is not an aten relu operator")
-
-            if _is_annotated([relu_node, conv_node]):
-                continue
-
-            input_qspec_map = {}
-            input_act = conv_node.args[0]
-            assert isinstance(input_act, Node)
-            input_qspec_map[input_act] = get_input_act_qspec(quantization_config)
-
-            weight = conv_node.args[1]
-            assert isinstance(weight, Node)
-            input_qspec_map[weight] = get_weight_qspec(quantization_config)
-
-            bias = conv_node.args[2]
-            if isinstance(bias, Node):
-                input_qspec_map[bias] = get_bias_qspec(quantization_config)
-
-            conv_node.meta["quantization_annotation"] = QuantizationAnnotation(
-                input_qspec_map=input_qspec_map, _annotated=True
-            )
-            relu_node.meta["quantization_annotation"] = QuantizationAnnotation(
-                output_qspec=get_output_act_qspec(quantization_config),  # type: ignore[arg-type]
-                _annotated=True,
-            )
+        return _OP_TO_ANNOTATOR["conv2d_relu"](gm, quantization_config, filter_fn)
 
     def _annotate_conv2d(
         self,
@@ -587,41 +418,7 @@ class XNNPACKQuantizer(Quantizer):
         quantization_config: QuantizationConfig,
         filter_fn: Optional[Callable[[Node], bool]] = None,
     ) -> None:
-        conv_partitions = get_source_partitions(
-            gm.graph, [torch.nn.Conv2d, torch.nn.functional.conv2d], filter_fn
-        )
-        conv_partitions = list(itertools.chain(*conv_partitions.values()))
-        for conv_partition in conv_partitions:
-            if len(conv_partition.output_nodes) > 1:
-                raise ValueError("conv partition has more than one output node")
-            conv_node = conv_partition.output_nodes[0]
-            if (
-                conv_node.op != "call_function"
-                or conv_node.target != torch.ops.aten.convolution.default
-            ):
-                raise ValueError(f"{conv_node} is not an aten conv2d operator")
-            # skip annotation if it is already annotated
-            if _is_annotated([conv_node]):
-                continue
-
-            input_qspec_map = {}
-            input_act = conv_node.args[0]
-            assert isinstance(input_act, Node)
-            input_qspec_map[input_act] = get_input_act_qspec(quantization_config)
-
-            weight = conv_node.args[1]
-            assert isinstance(weight, Node)
-            input_qspec_map[weight] = get_weight_qspec(quantization_config)
-
-            bias = conv_node.args[2]
-            if isinstance(bias, Node):
-                input_qspec_map[bias] = get_bias_qspec(quantization_config)
-
-            conv_node.meta["quantization_annotation"] = QuantizationAnnotation(
-                input_qspec_map=input_qspec_map,
-                output_qspec=get_output_act_qspec(quantization_config),
-                _annotated=True,
-            )
+        return _OP_TO_ANNOTATOR["conv2d"](gm, quantization_config, filter_fn)
 
     def _annotate_linear(
         self,
@@ -632,53 +429,13 @@ class XNNPACKQuantizer(Quantizer):
         return _OP_TO_ANNOTATOR["linear"](gm, quantization_config, filter_fn)
 
     # TODO: move this to BoltNNQuantizer?
-    def _annotate_gru(
+    def _annotate_gru_io_only(
         self,
         gm: torch.fx.GraphModule,
         quantization_config: QuantizationConfig,
         filter_fn: Optional[Callable[[Node], bool]] = None,
     ) -> None:
-        gru_partitions = get_source_partitions(gm.graph, [torch.nn.GRU], filter_fn)
-        gru_partitions = list(itertools.chain(*gru_partitions.values()))
-        for gru_partition in gru_partitions:
-            output_nodes = gru_partition.output_nodes
-            input_nodes = gru_partition.input_nodes
-            # skip annotation if it is already annotated
-            if _is_annotated(input_nodes + output_nodes):
-                continue
-            # inside each GRU partition, we should be able to annotate each linear
-            # subgraph
-            input_qspec_map: Dict[Node, QuantizationSpecBase] = {}
-            input_act = input_nodes[0]
-            input_act_user = list(input_act.users.keys())[0]
-            assert isinstance(input_act, Node)
-            assert isinstance(input_act_user, Node)
-            input_act_user.meta["quantization_annotation"] = QuantizationAnnotation(
-                input_qspec_map={
-                    input_act: get_input_act_qspec(quantization_config),
-                },
-                _annotated=True,
-            )
-
-            hidden_state = input_nodes[1]
-            hidden_state_user = list(hidden_state.users.keys())[0]
-            assert isinstance(hidden_state, Node)
-            assert isinstance(hidden_state_user, Node)
-            hidden_state_user.meta["quantization_annotation"] = QuantizationAnnotation(
-                input_qspec_map={
-                    hidden_state: get_input_act_qspec(quantization_config),
-                },
-                _annotated=True,
-            )
-
-            assert len(output_nodes) == 2, "expecting GRU to have two outputs"
-            for output in output_nodes:
-                output.meta["quantization_annotation"] = QuantizationAnnotation(
-                    output_qspec=get_output_act_qspec(quantization_config),
-                    _annotated=True,
-                )
-            nodes_to_mark_annotated = list(gru_partition.nodes)
-            _mark_nodes_as_annotated(nodes_to_mark_annotated)
+        return _OP_TO_ANNOTATOR["gru_io_only"](gm, quantization_config, filter_fn)
 
     def _annotate_maxpool2d(
         self,
@@ -686,43 +443,9 @@ class XNNPACKQuantizer(Quantizer):
         quantization_config: QuantizationConfig,
         filter_fn: Optional[Callable[[Node], bool]] = None,
     ) -> None:
-        module_partitions = get_source_partitions(
-            gm.graph, [torch.nn.MaxPool2d, torch.nn.functional.max_pool2d], filter_fn
-        )
-        maxpool_partitions = list(itertools.chain(*module_partitions.values()))
-        for maxpool_partition in maxpool_partitions:
-            output_node = maxpool_partition.output_nodes[0]
-            maxpool_node = None
-            for n in maxpool_partition.nodes:
-                if n.target == torch.ops.aten.max_pool2d_with_indices.default:
-                    maxpool_node = n
-            if _is_annotated([output_node, maxpool_node]):  # type: ignore[list-item]
-                continue
+        return _OP_TO_ANNOTATOR["maxpool2d"](gm, quantization_config, filter_fn)
 
-            input_act = maxpool_node.args[0]  # type: ignore[union-attr]
-            assert isinstance(input_act, Node)
-
-            # only annotate maxpool when the output of the input node is annotated
-            if (
-                "quantization_annotation" not in input_act.meta
-                or not input_act.meta["quantization_annotation"]._annotated
-                or input_act.meta["quantization_annotation"].output_qspec is None
-            ):
-                continue
-            # input and output of maxpool will share quantization parameter with input of maxpool
-            act_qspec = SharedQuantizationSpec(input_act)
-            # act_qspec = get_act_qspec(quantization_config)
-            maxpool_node.meta["quantization_annotation"] = QuantizationAnnotation(  # type: ignore[union-attr]
-                input_qspec_map={
-                    input_act: act_qspec,
-                },
-                _annotated=True,
-            )
-            output_node.meta["quantization_annotation"] = QuantizationAnnotation(
-                output_qspec=act_qspec,
-                _annotated=True,
-            )
-
+    # this should be handled by propagation, TODO: remove
     def _annotate_input_out_obs_sharing_op(
         self,
         op: Callable,
@@ -808,41 +531,7 @@ class XNNPACKQuantizer(Quantizer):
         quantization_config: QuantizationConfig,
         filter_fn: Optional[Callable[[Node], bool]] = None,
     ) -> None:
-        fused_partitions = find_sequential_partitions(
-            gm, [torch.add, torch.nn.ReLU], filter_fn
-        )
-        for fused_partition in fused_partitions:
-            add_partition, relu_partition = fused_partition
-            if len(relu_partition.output_nodes) > 1:
-                raise ValueError("Relu partition has more than one output node")
-            relu_node = relu_partition.output_nodes[0]
-            if len(add_partition.output_nodes) > 1:
-                raise ValueError("add partition has more than one output node")
-            add_node = add_partition.output_nodes[0]
-
-            if _is_annotated([relu_node, add_node]):
-                continue
-
-            input_act_qspec = get_input_act_qspec(quantization_config)
-            output_act_qspec = get_output_act_qspec(quantization_config)
-
-            input_qspec_map = {}
-            input_act0 = add_node.args[0]
-            if isinstance(input_act0, Node):
-                input_qspec_map[input_act0] = input_act_qspec
-
-            input_act1 = add_node.args[1]
-            if isinstance(input_act1, Node):
-                input_qspec_map[input_act1] = input_act_qspec
-
-            add_node.meta["quantization_annotation"] = QuantizationAnnotation(
-                input_qspec_map=input_qspec_map,
-                _annotated=True,
-            )
-            relu_node.meta["quantization_annotation"] = QuantizationAnnotation(
-                output_qspec=output_act_qspec,
-                _annotated=True,
-            )
+        return _OP_TO_ANNOTATOR["add_relu"](gm, quantization_config, filter_fn)
 
     def _annotate_add(
         self,
@@ -850,32 +539,7 @@ class XNNPACKQuantizer(Quantizer):
         quantization_config: QuantizationConfig,
         filter_fn: Optional[Callable[[Node], bool]] = None,
     ) -> None:
-        add_partitions = get_source_partitions(
-            gm.graph, [operator.add, torch.add], filter_fn
-        )
-        add_partitions = list(itertools.chain(*add_partitions.values()))
-        for add_partition in add_partitions:
-            add_node = add_partition.output_nodes[0]
-            if _is_annotated([add_node]):
-                continue
-
-            input_act_qspec = get_input_act_qspec(quantization_config)
-            output_act_qspec = get_output_act_qspec(quantization_config)
-
-            input_qspec_map = {}
-            input_act0 = add_node.args[0]
-            if isinstance(input_act0, Node):
-                input_qspec_map[input_act0] = input_act_qspec
-
-            input_act1 = add_node.args[1]
-            if isinstance(input_act1, Node):
-                input_qspec_map[input_act1] = input_act_qspec
-
-            add_node.meta["quantization_annotation"] = QuantizationAnnotation(
-                input_qspec_map=input_qspec_map,
-                output_qspec=output_act_qspec,
-                _annotated=True,
-            )
+        return _OP_TO_ANNOTATOR["add"](gm, quantization_config, filter_fn)
 
     def validate(self, model: torch.fx.GraphModule) -> None:
         pass
